requirements.txt
appdirs==1.4.3
asgiref==3.2.10
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
Django==3.1
django-filter==2.3.0
django-fullclean==0.0.4
djangorestframework==3.11.1
djangorestframework-filters==1.0.0.dev2
drf-yasg==1.17.1
html5lib==1.0.1
idna==2.8
inflection==0.5.1
install==1.3.3
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.14.0
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.16.11
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
